Delete a bogus attempt at supporting self-signed certificates
Apparently, this was a shameful misconception: if a way to verify
a host certificate without referring to CA exists, the -E isn't it.
So, people placing their clipboards at hosts with self-signed
certificates should use NOCERT for now (or convert those certificates
into trivial CA bundles somehow, if that is possible).